requirements.txt
asttokens==2.4.1
bleach==6.1.0
bokeh==3.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.2.0
cvopt==0.4.3
cvxopt==1.3.2
cycler==0.12.1
Cython==3.0.7
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
future==0.18.3
GPy==1.10.0
GPyOpt==1.2.6
hyperopt==0.2.7
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.6.1
kaggle==1.5.16
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.4
paramz==0.9.5
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==23.12.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
xyzservices==2023.10.1
